requirements.txt
aniso8601==9.0.1
astroid==1.6.6
attrs==21.2.0
behave==1.2.6
click==7.1.2
cloudant==2.14.0
colorama==0.4.4
compare==0.2b0
coverage==6.0.2
codecov==2.1.10
factory-boy==3.2.0
Faker==9.5.2
flake8==3.5.0
Flask==1.1.4
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
honcho==1.0.1
httpie==2.4.0
iniconfig==1.1.1
isort==5.9.3
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.0.0
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
nose==1.3.7
packaging==21.0
pinocchio==0.4.3
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.10.0
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.9.3
pyparsing==3.0.1
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
rednose==1.3.0
retry==0.9.2
requests==2.25.1
selenium==3.141.0
six==1.16.0
SQLAlchemy==1.3.24
termstyle==0.1.11
text-unidecode==1.3
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.13.2
Flask-RESTX==0.5.1
